feat: package library
