+++ package.json
@@ -2,15 +2,17 @@
   "name": "snyk-policy",
   "version": "0.0.0",
   "description": "Snyk's policy parser and matching logic",
-  "files": [
-    "dist/package.json",
-    "dist/lib",
-    "LICENSE",
-    "README.md"
-  ],
-  "main": "./dist/lib/index.js",
+  "type": "module",
+  "files": ["dist"],
+  "module": "./dist/index.js",
+  "types": "./dist/lib/index.d.ts",
+  "exports": {
+    ".": {
+      "import": "./dist/index.js"
+    }
+  },
   "scripts": {
-    "build": "npm run clean && tsc",
+    "build": "npm run clean && tsc && vite build",
     "clean": "rm -rf ./dist",
     "dev": "tsc -w",
     "env": "node -e 'console.log(process.env, process.versions)'",
@@ -42,6 +44,8 @@
     "prettier": "^2.0.5",
     "ts-node": "^10.9.1",
     "typescript": "~5.0.4",
+    "vite": "^4.3.9",
+    "vite-plugin-dts": "^2.3.0",
     "vitest": "^0.30.1"
   },
   "dependencies": {
